refactor: moved media upload logic to a custom hook
